% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asitis_init.R
\name{asitis_init}
\alias{asitis_init}
\title{Initialisation of Nunito font, Cairo bitmap printing type
You can use asitis_init() or just init()}
\usage{
asitis_init()
}
\description{
Initialisation of Nunito font, Cairo bitmap printing type
You can use asitis_init() or just init()
}
\examples{
asitis_init()
}
